app_django_pytest/serializers.py
from rest_framework import serializers

from app_django_pytest.models import Message


class MessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Message
        fields = ['id', 'user', 'text', 'created_at']